Fix server crash by exporting updateProfile controller
The server was failing to start with the error 'Route.put() requires a callback function but got a [object Undefined]' in authRoutes.js. This was caused by the updateProfile function not being exported from authController.js.

This commit adds updateProfile to the module.exports object in server/controllers/authController.js, resolving the undefined callback error.